Resolve "B - test pass 3 while et if then else"

Closes #66

See merge request leducb/projet-gl!28

+++ src/main/java/fr/ensimag/deca/tree/ReadInt.java
@@ -18,7 +18,7 @@ public class ReadInt extends AbstractReadExpr {
     @Override
     public Type verifyExpr(DecacCompiler compiler, EnvironmentExp localEnv,
             ClassDefinition currentClass) throws ContextualError {
-        return compiler.environmentType.FLOAT;
+        return compiler.environmentType.INT;
     }
 
     @Override
